Improved commenting in flatf.c; Changed index identifier for team_field_list functions from int to size_t

+++ src/include/ncrunch/ncrunch.h
@@ -52,10 +52,10 @@ struct team_field_list {
 };
 
 int team_field_list_create(size_t num_fields);
-int team_field_list_set_name(int id, const char *name);
-int team_field_list_set_type(int id, enum team_field_type type);
-const char *team_field_list_get_name(int id);
-enum team_field_type team_field_list_get_type(int id);
+int team_field_list_set_name(size_t id, const char *name);
+int team_field_list_set_type(size_t id, enum team_field_type type);
+const char *team_field_list_get_name(size_t id);
+enum team_field_type team_field_list_get_type(size_t id);
 
 
 /**
